First upload of XML
This is A.E.Waite’s text from the Key to the Tarot, marked up with an XML that helps NTLK Wordnet parsers understand how to interpret the florid language for lexical analysis: reducing complex phrases to core ideas, as well as specifying must-interpret-as and do-not-interpret-as. Note that this only includes detailed markup of the apodoses in the text and not the protases.